requirements.txt
Flask
pandas
numpy
scikit-learn
tensorflow
keras
joblib
mysql-connector-python
sqlalchemy